Refactor XML utils
